Lock 的被他人解锁的问题

+++ app/src/main/java/com/beyond/note5/sync/webdav/DavLock.java
@@ -5,6 +5,8 @@ import android.util.Log;
 import com.alibaba.fastjson.JSONObject;
 import com.beyond.note5.sync.webdav.client.DavClient;
 
+import org.apache.commons.lang3.StringUtils;
+
 import java.io.IOException;
 import java.util.Date;
 
@@ -14,22 +16,16 @@ public class DavLock implements Lock {
 
     private DavClient client;
 
-    public DavLock(DavClient client, String url) {
+    private String id;  // 防止其他 dataSource 释放自己的 lock
+
+    public DavLock(DavClient client, String url,String id) {
         this.url = url;
         this.client = client;
+        this.id = id;
     }
 
     public boolean tryLock() {
-        if (isLocked()) {
-            return false;
-        }
-        try {
-            client.put(url, "");
-            return true;
-        } catch (IOException e) {
-            e.printStackTrace();
-            return false;
-        }
+        return tryLock(-1L);
     }
 
     @Override
@@ -38,7 +34,7 @@ public class DavLock implements Lock {
             if (isLocked()) {
                 return false;
             }
-            String json = JSONObject.toJSONString(new LockTimeUnit(new Date(), time));
+            String json = JSONObject.toJSONString(new LockTimeUnit(new Date(), time,id));
             client.put(url, json);
             return true;
         }catch (Exception e){
@@ -63,8 +59,14 @@ public class DavLock implements Lock {
 
     public boolean release() {
         try {
-            client.delete(url);
-            return true;
+            String json = client.get(url);
+            LockTimeUnit lockTimeUnit = JSONObject.parseObject(json, LockTimeUnit.class);
+            if (StringUtils.equals(lockTimeUnit.getId(),id)){
+                client.delete(url);
+                return true;
+            }else {
+                return false;
+            }
         } catch (IOException e) {
             Log.e(getClass().getSimpleName(),"release fail",e);
             return false;
@@ -72,15 +74,17 @@ public class DavLock implements Lock {
     }
 
     private static class LockTimeUnit {
+        private String id;
         private Date lastLockTime;
         private Long lockPeriod;
 
         public LockTimeUnit() {
         }
 
-        LockTimeUnit(Date lastLockTime, Long lockPeriod) {
+        LockTimeUnit(Date lastLockTime, Long lockPeriod, String id) {
             this.lastLockTime = lastLockTime;
             this.lockPeriod = lockPeriod;
+            this.id = id;
         }
 
         public Date getLastLockTime() {
@@ -99,8 +103,16 @@ public class DavLock implements Lock {
             this.lockPeriod = lockPeriod;
         }
 
+        public String getId() {
+            return id;
+        }
+
+        public void setId(String id) {
+            this.id = id;
+        }
+
         boolean expired() {
-            return System.currentTimeMillis() - lastLockTime.getTime() > lockPeriod;
+            return lastLockTime.getTime() != -1L && System.currentTimeMillis() - lastLockTime.getTime() > lockPeriod;
         }
     }
 }
